Map popup to same card + Changed default zoom bc birds travel more
