+++ src/assets/css/custom.scss
@@ -46,6 +46,7 @@ label {
 .fa-map-marker-alt {
   font-size: 3rem;
   color: $primary;
+  cursor: pointer;
 }
 
 .card {
