+++ src/components/birds/BirdSightingCard.scss
@@ -1,6 +1,6 @@
 @import 'variables';
 
-.SightingCard {
+.BirdSightingCard {
   cursor: pointer;
 
   .card-body {
